Moved function calls out of the BOOST_VERIFY macro
Due to the structure of some of the "assert" macros used by BOOST_VERIFY, it is unsafe to pass it function calls that are not idempotent (and thus good practice to not pass it function calls in general). In this particular instance, the pthread_mutex_lock call would be duplicated on a non-zero return from pthread_mutex_lock, resulting in a possible double lock. This issue was uncovered with Coverity.